chore: set version of ethers to 2.0.2 & remove provider and transaction crates
