--- crates/transaction/src/lib.rs
@@ -1,39 +1,0 @@
-use ethers_core::{
-    types::{
-        transaction::eip2718::TypedTransaction, Address, Bytes,
-        TransactionRequest, U256, Eip1559TransactionRequest,
-    },
-};
-
-pub struct EthersTxOptions {
-    pub gas_limit: Option<U256>,
-    pub max_fee_per_gas: Option<U256>,
-    pub max_priority_fee_per_gas: Option<U256>,
-    pub gas_price: Option<U256>,
-    pub value: Option<U256>,
-    pub nonce: Option<U256>,
-}
-
-pub fn create_transaction(address: Option<Address>, data: Bytes, options: &EthersTxOptions) -> TypedTransaction {
-    if options.gas_price.is_some() {
-        return TransactionRequest {
-            to: address.map(Into::into),
-            data: Some(data),
-            gas: options.gas_limit,
-            gas_price: options.gas_price,
-            value: options.value,
-            nonce: options.nonce,
-            ..Default::default()
-        }.into();
-    }
-    Eip1559TransactionRequest {
-        to: address.map(Into::into),
-        data: Some(data),
-        gas: options.gas_limit,
-        max_fee_per_gas: options.max_fee_per_gas,
-        max_priority_fee_per_gas: options.max_priority_fee_per_gas,
-        value: options.value,
-        nonce: options.nonce,
-        ..Default::default()
-    }.into()
-}
