+++ crates/util/Cargo.toml
@@ -12,8 +12,8 @@ edition = "2021"
 
 [dependencies]
 thiserror = "1.0.38"
-polywrap-wasm-rs = "0.10.0-pre.10"
+polywrap-wasm-rs = "0.10.2"
 serde = { version = "1.0.152", features = ["derive"] }
 serde_json = { version = "1.0.93", default-features = false, features = ["raw_value"] }
-ethers-core = { version="2.0.3", features = ["eip712"] }
+ethers-core = { version="=2.0.2", features = [] }
 hex = { version = "0.4.3", default-features = false, features = ["alloc"] }
