+++ wraps/core/Cargo.toml
@@ -15,12 +15,11 @@ thiserror = "1.0.38"
 polywrap-wasm-rs = "0.10.0"
 serde = { version = "1.0.152", features = ["derive"] }
 serde_json = { version = "1.0.93", default-features = false, features = ["raw_value"] }
-ethers-providers = { version = "=2.0.3", features = [] }
-ethers-core = { version="2.0.3", features = ["eip712"] }
-ethers-signers = { version = "=2.0.3", features = [] }
+ethers-providers = { version = "=2.0.2", features = [] }
+ethers-core = { version="=2.0.2", features = [] }
+ethers-signers = { version = "=2.0.2", features = [] }
 hex = { version = "0.4.3", default-features = false, features = ["alloc"] }
 ethers-utils = { path = "../../crates/util" }
-ethers-provider = { path = "../../crates/provider" }
 
 [lib]
 crate-type = ["cdylib", "rlib"]
