+++ wraps/core/polywrap.build.yaml
@@ -1,12 +1,4 @@
 format: 0.2.0
-strategies:
-  image:
-    node_version: "17.9.1"
-    include:
-      - ./src
-      - ./package.json
 linked_packages:
   - name: "ethers-utils"
     path: ../../crates/util
-  - name: "ethers-provider"
-    path: ../../crates/provider
