+++ wraps/utils/Cargo.toml
@@ -14,7 +14,7 @@ edition = "2021"
 thiserror = "1.0.38"
 polywrap-wasm-rs = "0.10.0"
 serde = { version = "1.0.152", features = ["derive"] }
-ethers-core = { version="2.0.3", features = ["eip712"] }
+ethers-core = { version="=2.0.2" }
 hex = { version = "0.4.3", default-features = false, features = ["alloc"] }
 ethers-utils = { path = "../../crates/util" }
